requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.9
django-appconf==1.0.2
django-filter==1.0.2
django-user-accounts==2.0.0
djangorestframework==3.2.3
djangorestframework-jwt==1.10.0
drfdocs==0.0.11
Markdown==2.6.8
oauthlib==2.0.2
packaging==16.8
PyJWT==1.4.2
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
